Fix to make work with all stores
Not all stores are side-effecty, so in order for the data to load here we need to assign from the return value of store.get. If this isn't there then the tiddler data passed down the pipe just title, bag and a _uri field which means that things like tag subscriptions in `twsock` do not work.